+++ src/App.jsx
@@ -3,19 +3,20 @@ import { useEffect } from "react";
 import { Helmet, HelmetProvider } from "react-helmet-async";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./App.css";
-import Footer from "./components/Footer/Footer";
-import NavBar from "./components/NavBar/NavBar";
+import "./styles/bootstrap.css";
+import Footer from "./modules/components/Footer/Footer.jsx";
+import NavBar from "./modules/components/NavBar/NavBar.jsx";
 import { QuoteProvider } from "./modules/Quote/context/QuoteContext.jsx";
 import Contact from "./pages/Contact/Contact";
 import FAQ from "./pages/FAQ/FAQ";
-import HomePage2 from "./pages/HomePage2/HomePage2.jsx";
 import IndividualMaterial from "./pages/IndividualMaterial/IndividualMaterial.jsx";
-import Materials from "./pages/Materials/Materials.jsx";
+import Materials from "./modules/Materials/pages/Materials/Materials.jsx";
 import NotFound from "./pages/NotFound/NotFound";
 
 import p1sFarm from "./assets/homePage/p1s-farm.png";
 import { QuoteRouter } from "./modules/Quote/Quote.router.jsx";
 import { Box, Container } from "@mui/material";
+import Home from "./modules/Home/pages/Home.jsx";
 
 function App() {
   useEffect(() => {
@@ -38,21 +39,20 @@ function App() {
         <BrowserRouter>
           <NavBar />
           <div className="utility-wrapper">
-          {/* <Container maxWidth="md" sx={{ mt: 14 }}> */}
-              <Routes>
-                <Route path="/" element={<HomePage2 />} />
-                <Route path="/home" element={<HomePage2 />} />
-                <Route path="/contacto" element={<Contact />} />
-                <Route path="/not-found" element={<NotFound />} />
-                <Route path="/materiales" element={<Materials />} />
-                <Route
-                  path="/materiales/:material"
-                  element={<IndividualMaterial />}
-                />
-                <Route path="/preguntas-frecuentes" element={<FAQ />} />
-                <Route path="/cotizador/*" element={<QuoteRouter />} />
-              </Routes>
-          {/* </Container> */}
+            <Routes>
+              <Route path="/" element={<Home />} />
+              <Route path="/home" element={<Home />} />
+              <Route path="/contact" element={<Contact />} />
+              <Route path="/contacto" element={<Contact />} />
+              <Route path="/not-found" element={<NotFound />} />
+              <Route path="/materiales" element={<Materials />} />
+              <Route
+                path="/materiales/:material"
+                element={<IndividualMaterial />}
+              />
+              <Route path="/preguntas-frecuentes" element={<FAQ />} />
+              <Route path="/cotizador/*" element={<QuoteRouter />} />
+            </Routes>
           </div>
           <Footer />
         </BrowserRouter>
