--- src/components/Gallery/Gallery.css
@@ -1,63 +1,0 @@
-.gallery{
-    /* margin-top: 100px; */
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 40px;
-}
-
-.gallery_title{
-    font-size: 48px;
-    margin: 0 0 15px 0;
-    font-weight: 600;
-    line-height: 1.25;
-    letter-spacing: -.02em;
-    color: rgba(var(--text_color), 0.8);
-}
-
-.gallery_container{
-    width: 1400px;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    align-items: center;
-    gap: 50px;
-}
-
-.gallery_item{
-    width: 400px;
-    height: 400px;
-    border-radius: 10px;
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-    overflow: hidden;
-}
-
-.gallery_item img{
-    transition: 0.5s;
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    border-radius: 10px;
-}
-
-.gallery_item img:hover{
-    transition: 0.5s;
-    transform: scale(1.2);
-}
-
-@media (max-width: 440px) {
-    .gallery{
-        width: 100%;
-    }
-
-    .gallery_container{
-        width: calc(100vw - 40px);
-        flex-direction: column;
-    }
-
-    .gallery_item{
-        width: calc(100% - 40px);
-        height: auto;
-    }
-}
