--- src/components/MaterialDisplay/MaterialDisplay.css
@@ -1,290 +1,0 @@
-.material_display{
-    width: 100%;
-    margin-top: 50px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 40px;
-    font-weight: 700;
-    line-height: 1.2;
-    font-size: 40px;
-    color: var(--text_color);
-    /* margin-bottom: 200px; */
-}
-
-.material_display_container{
-    /* box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px; */
-    /* padding: 40px; */
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 15px;
-    gap: 50px;
-}
-
-.material_display_item{
-    min-width: 100%;
-    display: grid;
-    grid-template-columns: 300px 600px;
-    gap: 50px;
-    justify-self: center;
-    margin-top: 30px;
-    justify-content: center;
-    align-items: center;
-}
-
-.material_display_item img{
-    width: 300px;
-    height: 300px;
-    justify-self: center;
-    /* background: #f0f1f8; */
-    border-radius: 8px;
-}
-
-.material_display_item_container{
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-}
-
-.material_display_name{
-    font-size: 30px;
-    font-weight: 700;
-    color: rgb(var(--text_color));
-}
-
-.material_display_description{
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 1.5;
-    color: rgb(var(--text_color));
-}
-
-.material_display_type{
-    background: rgba(var(--primary_color), 0.2);
-    /* background: #f0f1f8; */
-    width: fit-content;
-    padding: 5px 10px;
-    border-radius: 15px;
-    font-size: 18px;
-    line-height: 1.5;
-    font-weight: 700;
-}
-
-.material_display_extra{
-    font-size: 16px;
-    color: #747892;
-}
-
-.material_display_select{
-    transition: 0.5s;
-    background-color: rgb(var(--primary_color));
-    color: var(--background_color);
-    text-align: center;
-    padding: 10px;
-    border-radius: 8px;
-    font-size: 16px;
-    font-weight: 400;
-    margin-top: 20px;
-}
-
-.material_display_select:hover{
-    transition: 0.5s;
-    background-color: rgba(var(--primary_color), 0.8);
-}
-
-.material_display_colors{
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.material_display_colors div{
-    width: 10px;
-    height: 10px;
-    background-color: #fff;
-    border: 0.1px solid #000;
-    border-radius: 100%;
-}
-
-.material_display_colors_select{
-    transition: 0.5s;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: start;
-    justify-content: start;
-    gap: 10px;
-}
-
-.material_display_infills_button{
-    transition: 0.5s;
-    border-radius: 8px;
-    font-size: 18px;
-    width: fit-content;
-    line-height: 1.5;
-    font-weight: 700;
-    padding: 5px 10px;
-    background-color: rgba(var(--primary_color), 0.15);
-    color: #747892;
-}
-
-.material_display_infills_button:hover{
-    transition: 0.5s;
-    background-color: rgba(var(--primary_color), 0.3);
-}
-
-.selected_standard{
-    transition: 0.5s;
-    border-radius: 8px;
-    font-size: 18px;
-    width: fit-content;
-    line-height: 1.5;
-    font-weight: 700;
-    padding: 5px 10px;
-    background-color: rgba(var(--primary_color), 0.35);
-    color: #747892;
-}
-
-.material_display_colors_select_button{
-    transition: 0.5s;
-    width: 30px;
-    height: 30px;
-    background-color: #fff;
-    border-radius: 100%;
-}
-
-.material_display_infills{
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.selected_color_standard{
-    width: 30px;
-    height: 30px;
-    border: 3px solid rgba(var(--primary_color), 1);
-    border-radius: 100%;
-}
-
-.material_display_progress{
-    font-size: 20px;
-    font-weight: 400;
-    color: rgba(var(--primary_color), 0.5);
-    text-align: left;
-    align-self: center;
-    width: 950px;
-}
-
-.material_display_progress a{
-    text-decoration: none;
-    color: rgba(var(--primary_color), 0.5);
-}
-
-.specs_description{
-    width: 950px;
-    text-align: left;
-    font-size: 30px;
-    font-weight: 300;
-}
-
-.specs_display{
-    width: 950px;
-    font-size: 20px;
-    font-weight: 400;
-    margin-top: 20px;
-    display: flex;
-    gap: 10px;
-}
-
-.specs_table{
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-}
-
-.specs_li{
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    padding: 10px;
-}
-
-.specs_li_title{
-    width: 100%;
-    text-align: center;
-    background-color: #6a6a6a;
-    color: #fff;
-    padding: 10px;
-    font-weight: 700;
-}
-
-@media (max-width: 440px) {
-
-    .material_display{
-        font-size: 30px;
-    }
-
-    #material_display_title{
-        width: 80vw;
-        text-align: center;
-    }
-
-    .material_display_progress{
-        width: fit-content;
-    }
-
-    .material_display_item{
-        display: flex;
-        flex-direction: column;
-        width: 300px;
-        gap: 40px;
-    }
-
-    .material_display_item img{
-        align-self: center;
-    }
-
-    .material_display_item_container{
-        padding: 0 30px;
-    }
-
-    .material_display_container{
-        box-shadow: none;
-        padding: 0;
-    }
-
-    .material_display_infills{
-        flex-wrap: wrap;
-    }
-
-    .specs_display{
-        flex-direction: column;
-        width: 100%;
-        justify-content: center;
-        align-items: center;
-        gap: 40px;
-    }
-
-    .specs_description{
-        width: 100%;
-        padding: 0 30px;
-        font-size: 22px;
-    }
-
-    .specs_table{
-        width: 90%;
-    }
-
-    .specs_li_title{
-        font-size: 18px;
-    }
-
-    .specs_li{
-        font-size: 14px;
-    }
-}
