--- src/components/NavBar/NavBar.css
@@ -1,145 +1,0 @@
-.nav_bar{
-    /* width: calc(100% - 80px);   es 100% menos el padding x 2 */
-    width: 100%;
-    height: var(--nav_bar_height);
-    position: fixed;
-    background-color: rgba(255, 255, 255, 0.9);
-    /* color: rgb(var(--primary_color)); */
-    font-size: 22px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0 40px;
-    top: 0;
-    z-index: 20;
-}
-
-.nav_bar_container{
-    /* min-width: 1400px; */
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.nav_bar_left{
-    padding: 0 30px 0 0;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    font-size: 40px;
-    text-decoration: none;
-    color: rgba(var(--text_color), 1);
-    font-weight: 700;
-}
-
-.nav_bar_left img{
-    width: 60px;
-    height: 60px;
-}
-
-.nav_bar_center{
-    display: flex;
-    justify-content: end;
-    align-items: center;
-    padding: 0 30px;
-    height: 100%;
-    font-size: 20px;
-    text-decoration: none;
-    color: rgba(var(--text_color), 1);
-    font-weight: 700;
-}
-
-.nav_bar_right{
-    padding: 0 0 0 30px;
-    height: 100%;
-    display: flex;
-    justify-content: start;
-    align-items: center;
-    gap: 15px;
-    font-size: 20px;
-    text-decoration: none;
-    color: rgba(var(--text_color), 1);
-    font-weight: 700;
-}
-
-.nav_bar_right a{
-    transition: 0.5s;
-    text-decoration: none;
-    color: rgba(var(--text_color), 0.6);
-    border: 1px solid rgba(var(--text_color), 0.6);
-    padding: 5px 15px;
-    font-weight: 400;
-    border-radius: 10px;
-    text-wrap: nowrap;
-}
-
-.nav_bar_right a:hover{
-    transition: 0.5s;
-    color: rgba(var(--text_color), 0.9);
-    border: 1px solid rgba(var(--text_color), 0.9);
-}
-
-#nav_bar_print{
-    color: var(--background_color);
-    background-color: rgba(var(--primary_color), 1);
-    border: 1px solid rgba(var(--primary_color), 1);
-    animation: shrinkButtonNavBar 2s infinite;
-}
-
-#nav_bar_print:hover{
-    background-color: rgba(var(--primary_color), 0.8);
-}
-
-@media (max-width: 440px) {
-    .nav_bar {
-        width: 100vw;
-        justify-content: space-between;
-        padding: 0;
-    }
-
-    .nav_bar_container{
-        width: 100%;
-        min-width: auto;
-    }
-
-    .nav_bar_left{
-        font-size: 18px;
-        text-wrap: nowrap;
-        justify-self: start;
-        align-items: start;
-        padding: 10px;
-    }
-
-    .nav_bar_left span{
-        display: none;
-    }
-
-    .nav_bar_right{
-        padding: 0;
-        padding-right: 20px;
-    }
-
-    .nav_bar_right a{
-        font-size: 14px;
-        height: 40px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        text-wrap: nowrap;
-    }
-}
-
-@keyframes shrinkButtonNavBar {
-
-    0% {
-        transform: translateX(2px) rotate(2deg);
-    }
-    50% {
-        transform: translateX(-2px) rotate(-2deg);
-    }
-    100% {
-        transform: translateX(2px) rotate(2deg);
-    }
-}
