--- src/components/WhoWeAre/WhoWeAre.css
@@ -1,110 +1,0 @@
-.who_we_are{
-    height: 100vh;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin-top: 100px;
-    margin-bottom: 100px;
-    /* border-radius: 5px; */
-}
-
-.who_we_are_container{
-    position: relative;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    gap: 100px;
-}
-
-.who_we_are_img{
-    position: absolute;
-    transition: 0.5s;
-    width: 100%;
-    /* height: 100%; */
-    overflow: hidden;
-    /* border-radius: 10px; */
-    object-fit: contain;
-    filter: brightness(0.7);
-}
-
-.who_we_are_img img{
-    transition: 0.5s;
-    width: auto;
-    /* height: calc(100% - 200px); */
-    height: 100%;
-    width: 100%;
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-}
-
-/* .who_we_are_img img:hover{
-    transition: 0.5s;
-    transform: scale(1.1);
-} */
-
-.who_we_are_presentation{
-    position: absolute;
-    z-index: 1;
-    width: 100%;
-    width: 1000px;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: 30px;
-}
-
-.who_we_are_title{
-    text-align: left;
-    font-size: 60px;
-    line-height: 1.5;
-    color: rgba(255, 255, 255, 0.9);
-}
-
-.who_we_are_text{
-    font-size: 28px;
-    line-height: 1.5;
-    color: rgba(255, 255, 255, 0.9);
-}
-
-@media (max-width: 440px) {
-    .who_we_are{
-        /* padding: 40px 20px; */
-        /* height: auto; */
-        /* width: auto; */
-        border-radius: 0;
-        margin-top: 100px;
-        margin-bottom: 100px;
-        height: 800px;
-    }
-
-    .who_we_are_container{
-        width: 100%;
-        height: 100%;
-        flex-direction: column;
-        gap: 40px;
-    }
-
-    .who_we_are_presentation{
-        width: 100%;
-        padding: 50px;
-    }
-
-    .who_we_are_title{
-        font-size: 40px;
-        color: rgba(var(--text_color), 0.8);
-    }
-
-    .who_we_are_text{
-        font-size: 22px;
-        text-align: center;
-        color: rgba(var(--text_color), 0.8);
-    }
-
-    .who_we_are_img{
-        display: none;
-    }
-}
