--- src/components/WhySection/WhySection.jsx
@@ -1,36 +1,0 @@
-import './WhySection.css';
-
-import impresion3dLogo from '../../assets/impresion-3d-logo.png';
-
-function WhySection(){
-    return(
-        <div className='why_section'>
-            <div className='why_section_left'>
-                {/* <span>Why Us</span> */}
-                <img src={impresion3dLogo}/>
-            </div>
-            <ol className='why_section_list'>
-                <li className='why_section_list_item'>
-                    <div className='why_section_list_item_content'>
-                        <span className='why_section_list_item_title'>Entregas rápidas y seguras</span>
-                        <span className='why_section_list_item_description'>Tu pedido será creado y entregado de la manera más profesional y eficiente posible.</span>
-                    </div>
-                </li>
-                <li className='why_section_list_item'>
-                    <div className='why_section_list_item_content'>
-                        <span className='why_section_list_item_title'>Servicio personalizado</span>
-                        <span className='why_section_list_item_description'>En la sección de comentarios puedes agregar cualquier solicitud a tu gusto para tus piezas.</span>
-                    </div>
-                </li>
-                <li className='why_section_list_item'>
-                    <div className='why_section_list_item_content'>
-                        <span className='why_section_list_item_title'>Gran soporte</span>
-                        <span className='why_section_list_item_description'>Contamos con una línea de chat de messenger para poder contestar todas tus dudas relacionadas a tu pedido.</span>
-                    </div>
-                </li>
-            </ol>
-        </div>
-    )
-}
-
-export default WhySection;
