+++ src/modules/Home/components/Gallery/Gallery.css
@@ -1,0 +1,41 @@
+.gallery{
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+    width: 100%;
+}
+
+.gallery_item{
+    border-radius: 10px;
+    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+    overflow: hidden;
+}
+
+.gallery_item img{
+    transition: 0.5s;
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+    border-radius: 10px;
+}
+
+.gallery_item img:hover{
+    transition: 0.5s;
+    transform: scale(1.2);
+}
+
+@media (max-width: 440px) {
+    .gallery{
+        width: 100%;
+    }
+
+    .gallery_container{
+        width: calc(100vw - 40px);
+        flex-direction: column;
+    }
+
+    .gallery_item{
+        width: calc(100% - 40px);
+        height: auto;
+    }
+}
