+++ src/modules/Materials/pages/Materials/Materials.css
@@ -1,0 +1,216 @@
+:root {
+  --card-height: 400px;
+  --card-width: calc(var(--card-height) / 1.5);
+  --card-border-radius: 10px;
+}
+
+.materials {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 90vh;
+  width: 100%;
+}
+
+.materials_page_display {
+  display: grid;
+  grid-template-columns: repeat(6, 1fr);
+  gap: 20px;
+  overflow: scroll;
+}
+
+::-webkit-scrollbar {
+  width: 10px;
+  height: 15px;
+}
+
+::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+::-webkit-scrollbar-thumb {
+  background: rgba(var(--text_color), 0.3);
+  border-radius: 10px;
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: rgba(var(--primary_color), 0.8);
+}
+
+.materials a {
+  text-decoration: none;
+  border-radius: var(--card-border-radius);
+}
+.materials_card {
+  width: var(--card-width);
+  height: var(--card-height);
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  padding: 0 36px;
+  perspective: 2500px;
+  /* margin: 0 30px; */
+  border-radius: var(--card-border-radius);
+}
+
+.materials_cover_image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: var(--card-border-radius);
+}
+
+.materials_wrapper {
+  transition: all 0.5s;
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  z-index: -1;
+  border-radius: var(--card-border-radius);
+  gap: 20px;
+}
+
+.materials_card:hover .wrapper {
+  transform: perspective(900px) translateY(-5%) rotateX(25deg) translateZ(0);
+  box-shadow: 2px 35px 32px -8px rgba(0, 0, 0, 0.75);
+  -webkit-box-shadow: 2px 35px 32px -8px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 2px 35px 32px -8px rgba(0, 0, 0, 0.75);
+}
+
+.materials_wrapper::before,
+.materials_wrapper::after {
+  content: "";
+  opacity: 0;
+  width: 100%;
+  height: 80px;
+  transition: all 0.5s;
+  position: absolute;
+  left: 0;
+  border-radius: var(--card-border-radius);
+}
+.materials_wrapper::before {
+  top: 0;
+  height: 100%;
+  background-image: linear-gradient(
+    to top,
+    transparent 46%,
+    rgba(12, 13, 19, 0.5) 68%,
+    rgba(12, 13, 19) 97%
+  );
+}
+.materials_wrapper::after {
+  bottom: 0;
+  opacity: 1;
+  background-image: linear-gradient(
+    to bottom,
+    transparent 46%,
+    rgba(12, 13, 19, 0.5) 68%,
+    rgba(12, 13, 19) 97%
+  );
+}
+
+.materials_card:hover .materials_wrapper::before,
+.materials_wrapper::after {
+  opacity: 0.8;
+}
+
+.materials_card:hover .materials_wrapper::after {
+  height: 120px;
+}
+.materials_title {
+  width: 100%;
+  line-height: 80px;
+  margin-bottom: 70px;
+  font-size: 60px;
+  text-align: center;
+  font-weight: 600;
+  text-decoration: none;
+  color: rgba(255, 255, 255, 0.9);
+  transition: transform 0.5s;
+}
+.materials_card:hover .materials_title {
+  transform: translate3d(0%, -40px, 100px);
+}
+
+.materials_character {
+  width: 100%;
+  opacity: 0;
+  transition: all 0.5s;
+  position: absolute;
+  z-index: -1;
+  border-radius: var(--card-border-radius);
+}
+
+.materials_card:hover .materials_character {
+  opacity: 1;
+  transform: translate3d(0%, -30%, 100px);
+  margin-bottom: 20px;
+}
+.materials_card:hover .materials_see_more {
+  opacity: 1;
+  transform: translate3d(0%, -30%, 100px);
+  margin-bottom: 20px;
+}
+
+.materials_see_more {
+  transition: 0.5s;
+  font-size: 44px;
+  opacity: 0;
+  text-align: center;
+  color: rgba(var(--secondary_color), 0.8);
+  transition: all 0.5s;
+  position: absolute;
+  z-index: -1;
+  background: linear-gradient(
+      to right,
+      rgba(var(--secondary_color), 0.2),
+      rgba(var(--primary_color), 0.5)
+    ),
+    linear-gradient(
+      to right,
+      rgba(var(--primary_color), 1),
+      rgba(var(--secondary_color), 1),
+      rgba(var(--primary_color), 1)
+    );
+  background-size: 100% 0.1em, 0 0.1em;
+  background-position: 100% 100%, 0 100%;
+  background-repeat: no-repeat;
+  transition: all 0.5s;
+}
+
+.materials_see_more:hover,
+.materials_see_more:focus {
+  transition: 0.5s;
+  /* text-decoration-line: underline; */
+  background-size: 0 0.1em, 100% 0.1em;
+}
+
+@media (max-width: 440px) {
+  :root {
+    --card-height: 500px;
+  }
+  .materials {
+    width: auto;
+  }
+  .materials_page_display {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    padding: 40px;
+    gap: 40px;
+    overflow: visible;
+  }
+  .materials_title {
+    font-size: 70px;
+  }
+  .materials_card:hover .materials_title {
+    transform: translate3d(0%, -7px, 100px);
+  }
+  .materials_card:hover .materials_character {
+    transform: translate3d(0%, -30%, 100px);
+  }
+  .materials_see_more {
+    font-size: 34px;
+  }
+}
