--- src/pages/HomePage/HomePage.css
@@ -1,16 +1,0 @@
-.home_page{
-    margin: var(--nav_bar_height) 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    /* gap: 80px; */
-}
-
-@media (max-width: 440px) {
-
-    .home_page{
-        gap: 0px;
-    }
-}
