--- src/pages/HomePage2/HomePage2.css
@@ -1,180 +1,0 @@
-.home{
-    margin: var(--nav_bar_height) 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    font-family: "League Spartan", sans-serif;
-}
-
-#home-title-separator{
-    font-weight: 600;
-    font-size: 80px;
-    text-align: center;
-    margin: 100px 0;
-}
-
-.home-separator-top{
-    width: 80vw;
-    height: 5px;
-    border-top: 2px dashed rgba(0, 0, 0, 0.4);
-    margin-top: 100px;
-}
-
-/* Carrusel */
-
-.home-carousel{
-    /* margin-top: 50px; */
-    width: 100%;
-}
-
-.carousel{
-    object-fit: cover;
-}
-
-.carousel-item{
-    width: 100px;
-    height: 800px;
-}
-
-.home-carousel-buttons{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 30px;
-    margin-bottom: 20px;
-}
-
-.home-carousel-buttons .btn{
-    font-size: 24px;
-}
-
-.home-img{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    filter: brightness(0.6);
-}
-
-.home-carousel h3{
-    font-size: 80px;
-}
-
-.home-carousel p{
-    font-size: 40px;
-}
-
-.contact-button{
-    animation: shrinkButtonNavBar 2s infinite;
-}
-
-/* Materiales */
-
-.home-materials{
-    width: 80%;
-}
-
-.card img{
-    width: 100%;
-    height: 300px;
-    transition: 1.5s;
-}
-
-.card img:hover{
-    transform: scale(1.05);
-    transition: 0.5s;
-}
-
-.card-img-top{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-/* Clientes */
-
-.home-clients{
-    height: fit-content;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    place-items: center;
-    gap: 100px;
-}
-
-.home-clients-img-container{
-    width: 300px;
-    height: 300px;
-    box-shadow: rgba(14, 30, 37, 0.12) 0px 2px 4px 0px, rgba(14, 30, 37, 0.32) 0px 2px 16px 0px;
-    border-radius: 10px;
-}
-
-.home-clients-img-container img{
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    border-radius: 10px;
-    filter: grayscale(0.4) contrast(0.6);
-    transition: 0.5s;
-}
-
-.home-clients-img-container img:hover{
-    filter: grayscale(0) contrast(1);
-    transition: 0.5s;
-}
-
-@media (max-width: 440px) {
-
-    .home-carousel{
-        width: 100%;
-    }
-    
-    .home-carousel h3{
-        font-size: 40px;
-    }
-
-    .home-carousel p{
-        font-size: 20px;
-    }
-
-    .home-carousel-buttons{
-        flex-direction: column;
-        margin-bottom: 40px;
-    }
-
-    .home-materials-container{
-        display: flex;
-        flex-direction: column;
-    }
-
-    #home-title-separator{
-        font-size: 60px;
-        margin: 60px;
-    }
-
-    .home-clients{
-        width: calc(100% - 40px);
-        display: flex;
-        flex-direction: column;
-
-    }
-
-    .home-clients-img-container{
-        width: 350px;
-        height: 350px;
-    }
-
-}
-
-@keyframes shrinkButtonNavBar {
-
-    0% {
-        transform: translateX(2px) rotate(2deg);
-    }
-    50% {
-        transform: translateX(-2px) rotate(-2deg);
-    }
-    100% {
-        transform: translateX(2px) rotate(2deg);
-    }
-}
