--- src/pages/HomePage2/responsive.css
@@ -1,111 +1,0 @@
-@media (max-width: 1300px) {
-  .footer_container {
-    border-radius: 250px 0 0 0;
-  }
-}
-
-@media (max-width: 1120px) {}
-
-@media (max-width: 992px) {
-  .hero_area {
-    height: auto;
-  }
-
-  .quote_btn-container {
-    display: none;
-  }
-
-  .custom_nav-container .navbar-nav {
-    margin: 0;
-    margin-left: auto;
-    align-items: center;
-  }
-
-  .custom_nav-container .navbar-nav {
-    padding-top: 15px;
-  }
-
-
-  .custom_nav-container .navbar-nav .nav-item .nav-link {
-    padding: 5px 25px;
-    margin: 5px 0;
-  }
-
-  .slider_section {
-    padding: 75px 0;
-  }
-
-  .contact_section .img-box {
-    margin-top: 45px;
-  }
-
-  .info_section .row>div:nth-child(1) {
-    margin-bottom: 25px;
-  }
-
-  .info_section .row>div:nth-child(2) {
-    margin-bottom: 25px;
-  }
-}
-
-@media (max-width: 768px) {
-  .hero_area {
-    border-radius: 0 0 250px 0;
-  }
-
-  .slider_section .detail-box {
-    margin-top: 45px;
-  }
-
-  .about_section {
-    border-radius: 150px 0 150px 0;
-  }
-
-  .about_section .img-box {
-    margin-top: 45px;
-  }
-
-  .contact_section .row {
-    flex-direction: column-reverse;
-  }
-
-  .contact_section .map_container {
-    margin-left: 0;
-    margin-bottom: 45px;
-    height: 375px;
-  }
-
-  .info_section .row>div:nth-child(3) {
-    margin-bottom: 35px;
-  }
-
-  .info_section .row>div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-
-  }
-
-  .info_section .info_links {
-    align-items: center;
-  }
-
-  .info_section form input {
-    text-align: center;
-  }
-}
-
-@media (max-width: 576px) {}
-
-@media (max-width: 480px) {}
-
-@media (max-width: 420px) {}
-
-@media (max-width: 376px) {}
-
-@media (min-width: 1200px) {
-  .container {
-    max-width: 1170px;
-  }
-}
